move downstreamChannelState into downstream
